Address comments from mirons.
Split up _base_repository into additional modules.

Make function returns consistent.
